+++ src/Model/Entity/Post.php
@@ -74,11 +74,10 @@ class Post extends Entity
             return null;
         }
 
-        //Checks for the first image in the text
-        preg_match('#<\s*img [^\>]*src\s*=\s*(["\'])(.*?)\1#im', $this->_properties['text'], $matches);
+        $preview = firstImageFromText($this->_properties['text']);
 
-        if (!empty($matches[2])) {
-            return Router::url($matches[2], true);
+        if ($preview) {
+            return $preview;
         }
 
         //Checks for a YouTube video and its preview
@@ -88,7 +87,7 @@ class Post extends Entity
             return Youtube::getPreview(isUrl($matches[1]) ? Youtube::getId($matches[1]) : $matches[1]);
         }
 
-        return null;
+        return false;
     }
 
     /**