+++ src/Model/Table/BannersPositionsTable.php
@@ -22,9 +22,6 @@
  */
 namespace MeCms\Model\Table;
 
-use Cake\ORM\Query;
-use Cake\ORM\RulesChecker;
-use MeCms\Model\Entity\BannersPosition;
 use MeCms\Model\Table\AppTable;
 
 /**
@@ -34,7 +31,7 @@ class BannersPositionsTable extends AppTable
 {
     /**
      * Name of the configuration to use for this table
-     * @var string|array
+     * @var string
      */
     public $cache = 'banners';