+++ tests/Fixture/BannersFixture.php
@@ -1,0 +1,95 @@
+<?php
+/**
+ * This file is part of MeTools.
+ *
+ * MeTools is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * MeTools is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with MeTools.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
+ * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
+ * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
+ * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ */
+namespace MeCms\Test\Fixture;
+
+use Cake\TestSuite\Fixture\TestFixture;
+
+/**
+ * BannersFixture
+ */
+class BannersFixture extends TestFixture
+{
+    /**
+     * Fields
+     * @var array
+     */
+    //@codingStandardsIgnoreStart
+    public $fields = [
+        'id' => ['type' => 'integer', 'length' => 11, 'unsigned' => false, 'null' => false, 'default' => null, 'comment' => '', 'autoIncrement' => true, 'precision' => null],
+        'position_id' => ['type' => 'integer', 'length' => 11, 'unsigned' => false, 'null' => false, 'default' => null, 'comment' => '', 'precision' => null, 'autoIncrement' => null],
+        'filename' => ['type' => 'string', 'length' => 255, 'null' => false, 'default' => null, 'collate' => 'latin1_swedish_ci', 'comment' => '', 'precision' => null, 'fixed' => null],
+        'target' => ['type' => 'string', 'length' => 255, 'null' => true, 'default' => null, 'collate' => 'latin1_swedish_ci', 'comment' => '', 'precision' => null, 'fixed' => null],
+        'description' => ['type' => 'string', 'length' => 255, 'null' => true, 'default' => null, 'collate' => 'latin1_swedish_ci', 'comment' => '', 'precision' => null, 'fixed' => null],
+        'active' => ['type' => 'boolean', 'length' => null, 'null' => false, 'default' => '1', 'comment' => '', 'precision' => null],
+        'click_count' => ['type' => 'integer', 'length' => 11, 'unsigned' => false, 'null' => false, 'default' => '0', 'comment' => '', 'precision' => null, 'autoIncrement' => null],
+        'created' => ['type' => 'datetime', 'length' => null, 'null' => true, 'default' => null, 'comment' => '', 'precision' => null],
+        'modified' => ['type' => 'datetime', 'length' => null, 'null' => true, 'default' => null, 'comment' => '', 'precision' => null],
+        '_indexes' => [
+            'position_id' => ['type' => 'index', 'columns' => ['position_id'], 'length' => []],
+        ],
+        '_constraints' => [
+            'primary' => ['type' => 'primary', 'columns' => ['id'], 'length' => []],
+        ],
+    ];
+    //@codingStandardsIgnoreEnd
+
+    /**
+     * Records
+     * @var array
+     */
+    public $records = [
+        [
+            'id' => 1,
+            'position_id' => 1,
+            'filename' => 'banner1.jpg',
+            'target' => '',
+            'description' => 'First banner',
+            'active' => 1,
+            'click_count' => 1,
+            'created' => '2016-12-26 16:26:04',
+            'modified' => '2016-12-26 16:26:04',
+        ],
+        [
+            'id' => 2,
+            'position_id' => 1,
+            'filename' => 'banner2.jpg',
+            'target' => '',
+            'description' => 'Second banner',
+            'active' => 0,
+            'click_count' => 0,
+            'created' => '2016-12-26 16:27:04',
+            'modified' => '2016-12-26 16:27:04',
+        ],
+        [
+            'id' => 3,
+            'position_id' => 2,
+            'filename' => 'banner3.jpg',
+            'target' => '',
+            'description' => 'Third banner',
+            'active' => 1,
+            'click_count' => 3,
+            'created' => '2016-12-26 16:28:04',
+            'modified' => '2016-12-26 16:28:04',
+        ],
+    ];
+}